Add files via upload
Two LED's had been used in the entire project . One is the internal led of Arduino connected to pin 13 and another is external led connected to pin 12 of Arduino
For the first two seconds an internal led will glow and an external led remains off.  After 2 seconds, the vice versa situation occurs  and switching b/w the two leds' go on in the same manner.